+++ srcs/cipher/des-cbc.c
@@ -2,7 +2,7 @@
 
 char			*des_cbc(unsigned char *str, size_t size, size_t *res_len, t_options *options)
 {
-	DPRINT("des_cbc(\"%.*s\", %zu)\n", size, str, size);
+	DPRINT("des_cbc(\"%.*s\", %zu)\n", (int)size, str, size);
 	(void)res_len;
 	(void)options;
 	(void)str;
